feat(sum): add keep_dim option to sum function
This commit introduces a new option to the sum function that allows
the user to keep the dimensionality of the input when summing over
a specific axis. This is useful in scenarios where the user wants to
retain the original shape of the data for further computations.

BREAKING CHANGE: The sum function now requires an additional boolean
argument, keep_dim. If set to true, the function will retain the
dimensionality of the input. If set to false, the function will behave
as before, reducing the dimensionality of the input by one.