feat(zenu-autograd): Implement sum function and add tests
This commit includes the implementation of the sum function in the
zenu-autograd library. It also includes the addition of tests for the sum
function. The sum function now supports 3D inputs and keeps dimensions if
specified. Debug and Display traits were also implemented for Variable.

In addition, the broadcast operation in the zenu-matrix library was updated
to support 4D inputs. Tests for the updated broadcast operation were added
as well.