refactor(zenu-matrix): move ones and zeros to constructor module
This commit moves the ones and zeros operations from the operation
module to a new constructor module in zenu-matrix. This change better
reflects the purpose of these operations as matrix constructors. All
references to these operations in zenu-matrix and zenu-autograd have
been updated to reflect this change. No functionality has been altered.