feat: Add trainable flag and related methods to Variable
This commit introduces a new flag `is_train` to the `Variable` struct
along with getter and setter methods. This flag is used to mark
variables that should be included in the training process.

A new method `get_all_trainable_variables` is also added which returns
all variables marked as trainable in the computational graph.

The `Matrix` struct is now `Hash` to allow it to be used in a `HashSet`.

The `Linear` layer now sets the `is_train` flag to true for its weights
and biases during initialization.

The `Model` trait's `forward` method is renamed to `predict`.

Tests are added to verify the functionality of the new methods.