feat: Add SGD optimizer and MSE loss function
This commit introduces the Stochastic Gradient Descent (SGD) optimizer
in the `zenu-optimizer` crate. The optimizer updates the parameters of
the layers based on their gradients and a learning rate.

A new loss function `mean_squared_error` is added to the
`zenu-autograd` crate. This function calculates the mean squared error
between two `Variable`s.

The `Model` trait in `zenu` crate is updated to include a new method
`layers` which returns a `Vec` of `Layer` trait objects.

The `Cargo.toml` file in the workspace root is updated to include
`zenu-optimizer` as a member of the workspace.

Tests are added to verify the functionality of the SGD optimizer and
the `mean_squared_error` function.