refactor(zenu): improve code quality and add features(clippy)
This commit refactors the zenu library to improve code quality and
add new features. It renames conv2d.rs to conv2d_impl.rs for clarity
and modifies reshape.rs and to_default_stride.rs to improve their
functionality. It also adds is_empty() methods to dataset.rs and
modifies the DataLoader iterator in the same file. Lastly, it
refactors mnist.rs to improve the download and extraction of the
MNIST dataset.