feat(conv2d): Add mutable reshape and fix gemm order
This commit introduces a mutable version of the reshape operation
in the matrix library. This allows reshaping a matrix in-place,
without creating a new one.

In addition, the order of arguments in the gemm operation inside
the Conv2dGrad function was incorrect. This has been fixed.

Lastly, a unit test for the conv2d operation has been added.