feat(conv2d): add bias to conv2d and deconv2d functions
This commit introduces bias to the conv2d and deconv2d functions. The
bias is optional and can be added when creating the Conv2d layer. The
bias is added to the result of the convolution operation. The backward
method of the Conv2d and Deconv2d functions has been updated to compute
the gradient of the bias. The tests for the conv2d and deconv2d
functions have been updated to include the bias.