feat(matrix): add serialization support for matrices
This commit introduces the ability to serialize and deserialize
matrices. This is done by adding the Serialize and Deserialize traits
from the serde library to the Matrix struct and its associated types.
This feature will allow matrices to be saved and loaded from disk,
which can be useful in various scenarios such as saving and loading
model weights in machine learning applications.

BREAKING CHANGE: The serde and serde_json dependencies have been added
to the zenu-matrix and zenu crates. If you are using these crates, you
will need to add these dependencies to your project as well.