feat(cublas): Add cublas_amax and cublas_dot functions
This commit introduces two new functions to the cublas module:
cublas_amax and cublas_dot. These functions provide the ability to
find the index of the maximum absolute value in a vector and to
calculate the dot product of two vectors, respectively. The
implementation supports both f32 and f64 data types. Unit tests
for these new functions are also included.